net: Rewrote getRemainingData
- Fixes read overflow error.

- Fixed temp_id on the state server being a uint16 instead of a uint32.

+++ src/stateserver/StateServer.cpp
@@ -63,7 +63,7 @@ void StateServer::handleGenerateInstanceObject(NetworkReader *reader)
 	uint16_t sender_pid = reader->readUint16();
 
 	// TempId.
-	uint16_t temp_id = reader->readUint32();
+	uint32_t temp_id = reader->readUint32();
 
 	// ParentId and ZoneId.
 	uint32_t parent_id = reader->readUint32();
